Remove newline from ruby version in tempfile name
The newline character at the end of the ruby version part of the temp
file name is causing odd behavior for me on OS X. In some cases it was
causing two actual files to be created. In others, it was causing the
newline to be replaced with `?`.

`chomp`ing the newline seems to resolve both of these issues.