Include require_relative RequireTree
Kernel.require_relative is actually defined in the load.c file of ruby's
source:

https://github.com/ruby/ruby/blob/b7ad549/load.c#L829-L846

Which effectively takes the current thread's callstack, determines
the caller's absolute filepath from that, and then runs the same
`rb_require_safe` as the C `require` function (`rb_f_require` is defined
right above the `require_relative` link from above).

Because we add another layer into the stack when monkey patching
`require_relative`, this doesn't work properly, and it tries to find the
file in the directory that require_relative was monkey patched from, not
from caller of `require_relative`.

The fix for this the monkey patching of `require_relative` was to call
into the monkey patched `require` defined above it, and pass it the
absolute path of the file based on the caller's absolute_path.  This is
similar to the C implementation, but does have the negative side effect
of probably not being as fast as it's C counter part, but I think that
this will be fine, as we are measuring memory, not boot time.