Don't use `before_filter` for test rails_app
Rails changed before_filter to become before_action, so prefer using
that over before_filter (since it has been removed in the newest
versions of Rails.

Also, since we are making `before_action` backwards compatible for older
versions of Rails where `before_filter` was the only method available,
move that alias builder to the ApplicationController instead of the
AuthenticatedController which was where it lived prior.