Use more descriptive commands and env vars
If you use a branch name with the method that is given i.e. "master" it
can mess things up since "master" is the commit SHA you specify in the
Gemfile.lock and not the actual master value (though this works
correctly if you're specifying path: in your Gemfile).

We can use SHA instead of BRANCHES for the input. Also just plain
"library" instead of "library_branches"