Allow `require_relative` to accept absolute paths
Ruby's `require_relative` [converts paths to absolute paths][1] before
calling `require`. This conversion effectively means that you can pass
an absolute path to `require_relative` and have it load the file
correctly, as if it were a relative path.

This is an idiosyncrasy of Ruby that doesn't match the semantics of the
method. However, some libraries, like [grpc][2], rely on the
idiosyncrasy to load files. Because Derailed's implementation doesn't
conform to this idiosyncrasy, it is unable to work when such libraries
exist in a project.

This change makes Derailed's implementation of `require_relative` behave
more closely to the implementation in Ruby. By making this change, you
can use Derailed with any offending gem.

[1]: https://github.com/ruby/ruby/blob/v2_6_3/load.c#L832-L841
[2]: https://github.com/grpc/grpc/blob/v1.22.0/src/ruby/lib/grpc/grpc.rb#L20