Don't use statistical significance as a stopping condition
The goal of adding the statistical significane check was to determine if it was "safe" to stop early, but this method of constantly checking for significance and then stopping produces less than stellar results. It's better to default to running the full test.

As a compromise the intermediate results are also output so if someone wants to use this as a quick and dirty guide for if they're on the right track or not, they can but they should know it might not hold up to scrutiny.